폴더 위치를 수정하라
주요 변경사항:
1. Contoller 폴더 내에 Domain과 Application 폴더를 옮겼습니다.
