+++ src/main/java/com/example/demo/domain/Location.java
@@ -1,4 +1,4 @@
-package com.example.demo.controllers.domain;
+package com.example.demo.domain;
 
 public record Location(
         String city,
